+++ modules/count.js
@@ -1,12 +1,12 @@
-import pokemons from "./pokemons.js";
+import pokemons from './pokemons.js';
 
 const pokemonCount = () => {
-    const total = document.querySelector('.nav-content');
-    total.innerHTML += `
+  const total = document.querySelector('.nav-content');
+  total.innerHTML += `
         <li id="pokemons" style="text-decoration: underline; font-weight: bold;">Pokemons(${pokemons.length})</li>
         <li>Nature</li>
         <li>Contest</li>`;
-    return pokemons.length;
-}
+  return pokemons.length;
+};
 
 export default pokemonCount;
