+++ src/index.html
@@ -1,32 +1,35 @@
 <!DOCTYPE html>
 <html lang="en">
-  <head>
-    <meta charset="UTF-8" />
-    <meta http-equiv="X-UA-Compatible" content="IE=edge" />
-    <meta name="viewport" content="width=device-width, initial-scale=1.0" />
-    <script src="https://kit.fontawesome.com/709d1a24ec.js" crossorigin="anonymous"></script>
-    <title>Pokeverse</title>
-  </head>
-  <body>
-    <header>
-      <nav class="my-nav">
-        <ul class="nav-ul">
-          <li class="nav-brand">
-            <p class="brand">POKEMON</p>
-          </li>
-          <li>
-            <ul class="nav-content">
-              <!-- list items of the nav -->
-            </ul>
-          </li>
-        </ul>
-      </nav>
-    </header>
-    <main>
-      <section id="container">
-        <!-- pokemon cards -->
-      </section>
-    </main>
-    <footer>Created by Microverse under CC license</footer>
-  </body>
+
+<head>
+  <meta charset="UTF-8" />
+  <meta http-equiv="X-UA-Compatible" content="IE=edge" />
+  <meta name="viewport" content="width=device-width, initial-scale=1.0" />
+  <script src="https://kit.fontawesome.com/709d1a24ec.js" crossorigin="anonymous"></script>
+  <title>Pokeverse</title>
+</head>
+
+<body>
+  <header>
+    <nav class="my-nav">
+      <ul class="nav-ul">
+        <li class="nav-brand">
+          <p class="brand">POKEMON</p>
+        </li>
+        <li>
+          <ul class="nav-content">
+            <!-- list items of the nav -->
+          </ul>
+        </li>
+      </ul>
+    </nav>
+  </header>
+  <main>
+    <section id="container">
+      <!-- pokemon cards -->
+    </section>
+  </main>
+  <footer>Created by Microverse under CC license</footer>
+</body>
+
 </html>
