+++ src/index.js
@@ -1,4 +1,4 @@
-import load from '../modules/api.js';
+import load from '../modules/load.js';
 import pokemonCount from '../modules/count.js';
 import { likeButton } from '../modules/like.js';
 import './style.css';
