+++ test/count.test.js
@@ -4,7 +4,7 @@ document.body.innerHTML = `
 <ul class="nav-content">
 
 </ul>
-`
+`;
 test('Check the pokemon count', () => {
   pokemonCount();
   const total = document.getElementById('pokemons');
